Non-parametric estimation of mean number of events

Usage

mean_no(
  formula,
  re_indicator,
  ce_indicator,
  data,
  re_control = list(),
  ce_control = list()
)

Arguments

formula

A formula passed to survfit.

re_indicator

The name of a variable indicating that these rows in the dataset belong to the risksets of the recurrent event process.

ce_indicator

The name of a variable indicating that these rows in the datasets belong to the riskset of the competing event process.

data

A data.frame in stacked format. The dataset needs to include one row for the competing event and one row for each risk episode of the recurrent event.

re_control

An optional list with arguments passed to survfit when computing risksets for the recurrent event.

ce_control

An optional list with arguments passed to survfit when computing risksets for the competing event.

Value

A data.frame including the estimated mean number of events expn at times t within strata strata.